+++ .changeset/add-multi-step-form-component.md
@@ -1,0 +1,5 @@
+---
+"@devopness/ui-react": minor
+---
+
+Added new `MultiStepForm` component in react ui library for navigating forms
