--- .changeset/add-step-form-component.md
@@ -1,5 +1,0 @@
----
-"@devopness/ui-react": minor
----
-
-Added new `StepForm` component in react ui library for navigating forms
