+++ packages/ui/react/src/components/Forms/MultiStepForm/index.ts
@@ -1,0 +1,1 @@
+export * from './MultiStepForm'
