--- packages/ui/react/src/components/Forms/StepForm/index.ts
@@ -1,1 +1,0 @@
-export * from './StepForm'
