+++ packages/ui/react/src/components/Forms/index.ts
@@ -9,4 +9,4 @@ export * from './Autocomplete'
 export * from './RadioSelectCards'
 export * from './Select'
 export * from './DynamicField'
-export * from './StepForm'
+export * from './MultiStepForm'
